Move data exporting methods to utils
